+++ src/logger.c
@@ -1995,11 +1995,12 @@ void disass(nes_state *state, char *output) {
     
     // BNE
   case 0xD0:
+      
     sprintf(output, "%04X  %02X %02X     BNE $%04X",
             state->cpu->current_opcode_PC,
             state->cpu->current_opcode,
             read_mem(state, state->cpu->current_opcode_PC+1),
-            state->cpu->current_opcode_PC + read_mem(state, state->cpu->current_opcode_PC+1) + 2);
+            state->cpu->current_opcode_PC + ((int8_t) read_mem(state, state->cpu->current_opcode_PC+1)) + 2);
     break;
 
     // CMP indirect-indexed,Y
